# swagger.yaml
openapi: 3.0.0
info:
  title: Mi API Final
  version: 1.0.0
  description: API RESTful con Express, MongoDB y JWT - Módulo 7
  contact:
    name: Proyecto Final - Módulo 7

servers:
  - url: http://localhost:3000
    description: Servidor local
  - url: https://mi-api-aventura-sigma.vercel.app
    description: Servidor en Vercel

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Ingresa el token JWT obtenido del endpoint /api/v1/login

paths:
  /:
    get:
      summary: Ruta de prueba
      description: Verifica que el servidor esté funcionando correctamente
      tags:
        - General
      responses:
        "200":
          description: Servidor funcionando
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: El servidor funciona correctamente

  /api/v1/hello:
    get:
      summary: Mensaje "Hola mundo"
      description: Devuelve un simple mensaje "Hola mundo"
      tags:
        - API v1
      responses:
        "200":
          description: Mensaje de bienvenida
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: Hola mundo

  /api/v1/saludo:
    get:
      summary: Saludo personalizado
      description: Devuelve un saludo personalizado con el nombre proporcionado
      tags:
        - API v1
      parameters:
        - in: query
          name: nombre
          required: true
          schema:
            type: string
          description: El nombre para personalizar el saludo
          example: Ana
      responses:
        "200":
          description: Saludo exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: Hola, Ana!
        "400":
          description: Parámetro nombre no proporcionado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Parámetro 'nombre' es requerido

  /api/v1/login:
    post:
      summary: Iniciar sesión (Autenticación)
      description: Genera un token JWT si las credenciales son correctas
      tags:
        - Autenticación
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: admin123
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        "400":
          description: Credenciales no proporcionadas
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Username y password son requeridos
        "401":
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Credenciales inválidas

  /api/v1/usuarios:
    get:
      summary: Obtener todos los usuarios
      description: Devuelve la lista de usuarios desde MongoDB (ruta protegida - requiere JWT)
      tags:
        - Usuarios
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de usuarios obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                      example: 507f1f77bcf86cd799439011
                    nombre:
                      type: string
                      example: Ana
                    email:
                      type: string
                      example: ana@example.com
                    edad:
                      type: number
                      example: 25
                    activo:
                      type: boolean
                      example: true
        "401":
          description: Token no proporcionado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Acceso denegado. Token no proporcionado.
        "403":
          description: Token inválido o expirado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Token inválido o expirado
        "500":
          description: Error del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Error al obtener usuarios

tags:
  - name: General
    description: Endpoints generales
  - name: API v1
    description: Endpoints de la versión 1 de la API
  - name: Autenticación
    description: Endpoints de autenticación y autorización
  - name: Usuarios
    description: Gestión de usuarios